# Conflicts:
#	sql/test.sql
#	src/main/resources/mapper/suggest/suggest.xml
#	src/main/webapp/WEB-INF/view/board/freeBoardUpdateForm.jsp
#	src/main/webapp/WEB-INF/view/user/kakaoSignup.jsp

+++ src/main/java/com/green/vrink/community/controller/FreeBoardRestController.java
@@ -79,7 +79,9 @@ public class FreeBoardRestController {
         Integer commUserId = freeBoardService.getUserId(freeBoardDTO.getCommunityId());
         Integer sessionUserId = user.getUserId();
 
-        if (userId != commUserId || sessionUserId != commUserId) {
+        if (userId.intValue() != commUserId.intValue()
+                || sessionUserId.intValue() != commUserId.intValue()){
+
             return new ResponseEntity<>("접근권한이 없습니다.", HttpStatus.FORBIDDEN);
         }
